requirements.txt
absl-py
jax[cpu]

matplotlib
numpy
setuptools
quantecon